backend/requirements.txt
Django>=5.0
psycopg2-binary>=2.9
django-cors-headers>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
gunicorn>=20.1
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0
drf-spectacular>=0.27.0